Fix DungeonView test and add color to MsgBox
Now dungeonview.py can be executed to have a small preview of its
content.

The message box display the last message in grey and the previous
messages in dark grey. The width was reduced to test the warping
which works just fine. So why bother with a MessageBox model
class. A simple list should be enough.